requirements.txt
aiogram>=3.0.0
requests>=2.31.0
aiosqlite>=0.19.0
aiofiles>=23.1.0
python-docx>=0.8.11
sentence-transformers>=2.2.2
numpy>=1.24.0
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.27.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
aiohttp>=3.9.0
psutil>=5.9.0 